Some minor changes in order to be able to better test the implementation.

+++ src/Benchmarking/Integer/IntegerCorectnessValidator.cpp
@@ -20,7 +20,7 @@ bool IntegerCorectnessValidator::validate(vector < long long unsigned int > & a,
 bool IntegerCorectnessValidator::validateVerbose(vector < long long unsigned int > & a, vector < long long unsigned int > & b) {
     unsigned int mmcnt = 0;
     for (unsigned int i = 0; i < a.size(); i++) {
-        if (a.at(i) != b.at(i)) {
+        if ((unsigned int) a.at(i) != (unsigned int) b.at(i)) {
             printf("Found mismatch at trip %u (indexing trips from 0).\n", i);
             printf("Vector 'a' contains: %llu, while vector 'b' contains: %llu.\n", a.at(i), b.at(i));
             mmcnt++;//return false;
